moved lifecycle demo to own component
