+++ src/app/lifecycle-demo/lifecycle-demo.component.html
@@ -1,0 +1,6 @@
+<div class="container p-5 text-center">
+  <button (click)="toggleSurprise()" type="button" class="btn btn-primary">
+    Surprise!
+  </button>
+  <hello *ngIf="surprise" [name]="'Tranglo'"></hello>
+</div>
